Function to find long lines in files
Outputs the filename, line number, and the part of the line which comes after the length limit

This function should work with any filename and any contents, but should not be parsed automatically. In order to make it completely safe you would have to use NUL as a separator instead of ":". This is left as an exercise to the reader.

This function also does not handle tabs specially, although that should be easy enough to fix for anyone who needs it.